Move 270/271 common sections to f27x.
